+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for 2.0.x
 ===================
 ## 2.1.0
 * [New] Support to extends TwigRenderizer with tags. See PR #65.
+* [New] Added `ItemSet` class to Core support classes.
 * [Improved] Improved permalinks customizations. See PR #64.
 
 ## 2.0.2 (2016-01-16)